✨ add expo-router, ui-kitten and structure folder
