+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "health-plaza-se-2-react-native",
   "version": "1.0.0",
-  "main": "expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -9,10 +9,21 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@eva-design/eva": "^2.2.0",
+    "@expo/metro-runtime": "^3.2.1",
+    "@ui-kitten/components": "^5.3.1",
     "expo": "~51.0.14",
+    "expo-constants": "^16.0.2",
+    "expo-linking": "^6.3.1",
+    "expo-router": "^3.5.16",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
-    "react-native": "0.74.2"
+    "react-dom": "^18.3.1",
+    "react-native": "0.74.2",
+    "react-native-safe-area-context": "^4.10.5",
+    "react-native-screens": "^3.32.0",
+    "react-native-svg": "9.13.6",
+    "react-native-web": "^0.19.12"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
