+++ src/app/about.tsx
@@ -1,0 +1,9 @@
+import { Text } from "react-native";
+
+const About = () => {
+  return (
+    <Text>about</Text>
+  )
+}
+
+export default About;
