+++ src/app/index.tsx
@@ -1,0 +1,9 @@
+import { Link } from 'expo-router';
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
+
+import HomePage from '@/modules/home/HomePage';
+
+export default function App() {
+  return <HomePage />;
+}
