+++ src/modules/quiz/QuizPage.tsx
@@ -1,0 +1,3 @@
+const QuizPage = () => {}
+
+export default QuizPage;
